Consolidate & refactor state initialization
* Put contracts dictionary initialization concerns into
contractStateUtils This includes, the refactored single exports from
generateContractInitialState.js and generateContractsInitialState.js and
getAbi.js

* remove unused module getAddress.js
