+++ src/generateStore.js
@@ -3,7 +3,7 @@ import { createStore, applyMiddleware, compose, combineReducers } from 'redux'
 import createSagaMiddleware from 'redux-saga'
 import drizzleSagas from './rootSaga'
 import drizzleReducers from './reducer'
-import { generateContractsInitialState } from './generateContractsInitialState'
+import { generateContractsInitialState } from './contractStateUtils'
 
 const composeSagas = sagas =>
   function * () {
@@ -24,7 +24,7 @@ const composeSagas = sagas =>
  * @returns {object} Redux store
  *
  */
-export function generateStore({
+export function generateStore ({
   drizzleOptions,
   appReducers = {},
   appSagas = [],
