--- src/getAddress.js
@@ -1,5 +1,0 @@
-export default function (contractConfig) {
-  if (contractConfig.web3Contract) {
-    return web3Contract.options.address
-  }
-}
